requirements.txt
django==3.2
djangorestframework==3.14.0
mysqlclient==2.0.1
